+++ gems.rb
@@ -23,7 +23,10 @@ group :test do
 	gem "decode"
 	
 	gem "rubocop"
+	gem "rubocop-md"
 	gem "rubocop-socketry"
+	
+	gem "sus-fixtures-async"
 	
 	gem "bake"
 	gem "bake-test"