+++ releases.md
@@ -1,5 +1,10 @@
 # Releases
 
+## Unreleased
+
+  - Rename `ContainerEnvironment` and `ContainerService` to `Managed::Environment` and `Managed::Service` respectively.
+  - Health check uses `Fiber.new{instance.ready!}.resume` to confirm fiber allocation is working.
+
 ## v0.14.4
 
   - Use `String::Format` gem for formatting.